+++ packages/angular/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.10](https://github.com/IBM/carbon-charts/compare/v0.7.9...v0.7.10) (2018-11-28)
+
+**Note:** Version bump only for package @carbon/charts-angular
+
+
+
+
+
+## [0.7.9](https://github.com/IBM/carbon-charts/compare/v0.7.8...v0.7.9) (2018-11-22)
+
+**Note:** Version bump only for package @carbon/charts-angular
+
+
+
+
+
+## [0.7.8](https://github.com/IBM/carbon-charts/compare/v0.7.7...v0.7.8) (2018-11-22)
+
+**Note:** Version bump only for package @carbon/charts-angular
+
+
+
+
+
+## [0.7.5](https://github.com/IBM/carbon-charts/compare/v0.7.4...v0.7.5) (2018-11-20)
+
+**Note:** Version bump only for package @carbon/charts-angular
+
+
+
+
+
+## [0.7.4](https://github.com/IBM/carbon-charts/compare/v0.7.3...v0.7.4) (2018-11-17)
+
+**Note:** Version bump only for package @carbon/charts-angular
+
+
+
+
+
 ## [0.7.3](https://github.com/IBM/carbon-charts/compare/v0.7.2...v0.7.3) (2018-11-13)
 
 **Note:** Version bump only for package @carbon/charts-angular